# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger bff-grpc",
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "paths": {
        "/mine/review/project/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "review"
                ],
                "summary": "获取项目列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "modeCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageNum",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "showStatus",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/stru.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pbs.ReviewProjectListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "pbs.ReviewProjectData": {
            "type": "object",
            "properties": {
                "created_date": {
                    "type": "string"
                },
                "created_user": {
                    "type": "string"
                },
                "deleted": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mode_code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updated_date": {
                    "type": "string"
                },
                "updated_user": {
                    "type": "string"
                }
            }
        },
        "pbs.ReviewProjectListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pbs.ReviewProjectData"
                    }
                },
                "page_num": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "description": "定义接收的参数\n参数类型 参数名 标识号(不可重复)",
                    "type": "integer"
                }
            }
        },
        "stru.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        }
    }
}